apps/backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Neo4j
neo4j==5.15.0

# Vector DB
qdrant-client==1.7.0

# Redis
redis==5.0.1
hiredis==2.2.3

# AI/ML
langchain==0.1.0
langgraph==0.0.20
langsmith==0.0.69
openai==1.6.1
anthropic==0.8.1
cohere==4.37

# Embeddings & Vector Operations
sentence-transformers==2.2.2
numpy==1.26.2

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Document Parsing
PyPDF2==3.0.1
python-docx==1.1.0

# Utils
python-dotenv==1.0.0
pydantic-core==2.14.5

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2
